requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
langchain_qdrant>=0.0.16
qdrant-client>=1.6.4
langchain_together>=0.0.3